misc: remove `default-starters.ts` and `dex-attributes.ts`
