+++ src/constants/game-constants.ts
@@ -4,6 +4,7 @@ import type { SystemSaveData } from "#types/system-data";
 /* biome-ignore-end lint/correctness/noUnusedImports: tsdoc imports */
 
 import { SpeciesFormKey } from "#enums/species-form-key";
+import { SpeciesId } from "#enums/species-id";
 import { StatusEffect } from "#enums/status-effect";
 
 /** Max value for an integer attribute in {@linkcode SystemSaveData} */
@@ -106,4 +107,35 @@ export const NON_VOLATILE_STATUS_EFFECTS = Object.freeze([
   StatusEffect.SLEEP,
   StatusEffect.FREEZE,
   StatusEffect.BURN,
+]);
+
+/** A list of Pokemon species that are unlocked when making a new account. */
+export const DEFAULT_STARTER_SPECIES = Object.freeze([
+  SpeciesId.BULBASAUR,
+  SpeciesId.CHARMANDER,
+  SpeciesId.SQUIRTLE,
+  SpeciesId.CHIKORITA,
+  SpeciesId.CYNDAQUIL,
+  SpeciesId.TOTODILE,
+  SpeciesId.TREECKO,
+  SpeciesId.TORCHIC,
+  SpeciesId.MUDKIP,
+  SpeciesId.TURTWIG,
+  SpeciesId.CHIMCHAR,
+  SpeciesId.PIPLUP,
+  SpeciesId.SNIVY,
+  SpeciesId.TEPIG,
+  SpeciesId.OSHAWOTT,
+  SpeciesId.CHESPIN,
+  SpeciesId.FENNEKIN,
+  SpeciesId.FROAKIE,
+  SpeciesId.ROWLET,
+  SpeciesId.LITTEN,
+  SpeciesId.POPPLIO,
+  SpeciesId.GROOKEY,
+  SpeciesId.SCORBUNNY,
+  SpeciesId.SOBBLE,
+  SpeciesId.SPRIGATITO,
+  SpeciesId.FUECOCO,
+  SpeciesId.QUAXLY,
 ]);
