+++ src/enums/ability-attr.ts
@@ -1,0 +1,7 @@
+/** Flags used to mark unlocked abilities of each species in dex data. */
+export const AbilityAttr = {
+  ABILITY_1: 1,
+  ABILITY_2: 2,
+  ABILITY_HIDDEN: 4,
+  PASSIVE: 8,
+} as const;
