+++ src/ui/handlers/game-stats-ui-handler.ts
@@ -1,8 +1,8 @@
 import { globalScene } from "#app/global-scene";
 import { GAME_HEIGHT, GAME_WIDTH } from "#constants/ui-constants";
-import { DexAttr } from "#data/dex-attributes";
 import { speciesStarterCosts } from "#data/starters";
 import { Button } from "#enums/button";
+import { DexAttr } from "#enums/dex-attr";
 import { TextStyle } from "#enums/text-style";
 import { UiMode } from "#enums/ui-mode";
 import type { GameData } from "#system/game-data";
