+++ src/ui/handlers/starter-select-ui-handler.ts
@@ -5,7 +5,6 @@ import { handleTutorial } from "#app/tutorial";
 import { PLAYER_PARTY_MAX_SIZE } from "#constants/game-constants";
 import { GAME_HEIGHT, GAME_WIDTH } from "#constants/ui-constants";
 import { allAbilities, allMoves, allSpecies } from "#data/data-lists";
-import { AbilityAttr, DexAttr } from "#data/dex-attributes";
 import { Egg, getEggTierForSpecies } from "#data/egg";
 import { speciesEggMoves } from "#data/egg-moves";
 import { getGrowthRateColor, getGrowthRateShadowColor } from "#data/exp";
@@ -29,10 +28,12 @@ import {
 } from "#data/starters";
 import type { Variant } from "#data/variant";
 import { getVariantTierForVariant, getVariantTint } from "#data/variant";
+import { AbilityAttr } from "#enums/ability-attr";
 import { AbilityId } from "#enums/ability-id";
 import { Button } from "#enums/button";
 import { ChallengeType } from "#enums/challenge-type";
 import { Device } from "#enums/device";
+import { DexAttr } from "#enums/dex-attr";
 import { DropDownColumn } from "#enums/drop-down-column";
 import { DropDownState } from "#enums/drop-down-state";
 import { DropDownType } from "#enums/drop-down-type";
